requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
transformers==4.1.1
seqeval==0.0.12
pytorch-crf==0.7.2
tensorflow==2.2.0
sentencepiece
tensorboard
numpy
tqdm